code cleanup and fix cornell with medium scene

+++ riow/vec3.h
@@ -113,7 +113,7 @@ vec3 reflect(const vec3& v, const vec3& n) {
 vec3 refract(const vec3& uv, const vec3& n, double etai_over_etat) {
     auto cos_theta = fmin(dot(-uv, n), 1.0);
     vec3 r_out_perp = etai_over_etat * (uv + cos_theta * n);
-    float sqlen = r_out_perp.length_squared();
+    double sqlen = r_out_perp.length_squared();
     vec3 r_out_parallel = sqlen >= 1.0f ? vec3(0, 0, 0) : -sqrt(1.0f - sqlen) * n;
     return r_out_perp + r_out_parallel;
 }
